# Conflicts:
#	.idea/assetWizardSettings.xml
#	.idea/caches/build_file_checksums.ser
#	.idea/codeStyles/codeStyleConfig.xml
#	.idea/misc.xml
#	app/src/main/java/ballidaku/mywallet/commonClasses/CommonMethods.java
#	app/src/main/java/ballidaku/mywallet/mainScreens/activities/AddBankDetails.java
#	app/src/main/java/ballidaku/mywallet/mainScreens/activities/MainActivity.java
#	app/src/main/java/ballidaku/mywallet/mainScreens/fragments/BankAccountsFragment.java
#	app/src/main/java/ballidaku/mywallet/mainScreens/fragments/MainFragment.java
#	app/src/main/java/ballidaku/mywallet/roomDatabase/AccountDetailsDataModelDao.java

+++ app/src/main/java/ballidaku/mywallet/mPin/PinLockAdapter.java
@@ -3,7 +3,7 @@ package ballidaku.mywallet.mPin;
 import android.content.Context;
 import android.graphics.PorterDuff;
 import android.graphics.Rect;
-import android.os.Build;
+import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.util.TypedValue;
@@ -12,6 +12,7 @@ import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
+import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 
@@ -25,6 +26,7 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
 
     private static final int VIEW_TYPE_NUMBER = 0;
     private static final int VIEW_TYPE_DELETE = 1;
+    private static final int VIEW_TYPE_RESET = 2;
 
     private Context mContext;
     private CustomizationOptionsBundle mCustomizationOptionsBundle;
@@ -35,14 +37,15 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
 
     private int[] mKeyValues;
 
-    public PinLockAdapter(Context context)
+    PinLockAdapter(Context context)
     {
         this.mContext = context;
         this.mKeyValues = getAdjustKeyValues(new int[]{1, 2, 3, 4, 5, 6, 7, 8, 9, 0});
     }
 
+    @NonNull
     @Override
-    public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType)
+    public RecyclerView.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType)
     {
         RecyclerView.ViewHolder viewHolder;
         LayoutInflater inflater = LayoutInflater.from(parent.getContext());
@@ -52,16 +55,21 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
             View view = inflater.inflate(R.layout.layout_number_item, parent, false);
             viewHolder = new NumberViewHolder(view);
         }
-        else
+        else if (viewType == VIEW_TYPE_DELETE)
         {
             View view = inflater.inflate(R.layout.layout_delete_item, parent, false);
             viewHolder = new DeleteViewHolder(view);
+        }
+        else
+        {
+            View view = inflater.inflate(R.layout.layout_delete_item, parent, false);
+            viewHolder = new ResetViewHolder(view);
         }
         return viewHolder;
     }
 
     @Override
-    public void onBindViewHolder(RecyclerView.ViewHolder holder, int position)
+    public void onBindViewHolder(@NonNull RecyclerView.ViewHolder holder, int position)
     {
 
         if (holder.getItemViewType() == VIEW_TYPE_NUMBER)
@@ -74,51 +82,63 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
             DeleteViewHolder vh2 = (DeleteViewHolder) holder;
             configureDeleteButtonHolder(vh2);
         }
+        else if (holder.getItemViewType() == VIEW_TYPE_RESET)
+        {
+            ResetViewHolder vh3 = (ResetViewHolder) holder;
+            configureResetButtonHolder(vh3);
+        }
     }
 
     private void configureNumberButtonHolder(NumberViewHolder holder, int position)
     {
         if (holder != null)
         {
-            if (position == 9)
-            {
-//                holder.mNumberButton.setVisibility(View.GONE);
-                holder.mNumberButton.setText("Reset");
-                holder.mNumberButton.setTag("reset");
-            }
-            else
-            {
-                holder.mNumberButton.setText(String.valueOf(mKeyValues[position]));
-                holder.mNumberButton.setVisibility(View.VISIBLE);
-                holder.mNumberButton.setTag(mKeyValues[position]);
-            }
+//            if (position == 9)
+//            {
+////                holder.mNumberButton.setVisibility(View.GONE);
+//                holder.mNumberButton.setText("Reset");
+//                holder.mNumberButton.setTag("reset");
+//            }
+//            else
+//            {
+            holder.mNumberButton.setText(String.valueOf(mKeyValues[position]));
+            holder.mNumberButton.setVisibility(View.VISIBLE);
+            holder.mNumberButton.setTag(mKeyValues[position]);
+//            }
 
             if (mCustomizationOptionsBundle != null)
             {
                 holder.mNumberButton.setTextColor(mCustomizationOptionsBundle.getTextColor());
                 if (mCustomizationOptionsBundle.getButtonBackgroundDrawable() != null)
                 {
-                    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.JELLY_BEAN)
-                    {
-                        holder.mNumberButton.setBackgroundDrawable(mCustomizationOptionsBundle.getButtonBackgroundDrawable());
-                    }
-                    else
-                    {
-                        holder.mNumberButton.setBackground(mCustomizationOptionsBundle.getButtonBackgroundDrawable());
-                    }
+                    holder.mNumberButton.setBackground(mCustomizationOptionsBundle.getButtonBackgroundDrawable());
                 }
 
                 holder.mNumberButton.setTextSize(TypedValue.COMPLEX_UNIT_PX, mCustomizationOptionsBundle.getTextSize());
 
 
 //                Log.e("POS ",""+position);
-                if (position == 9)
-                {
-//                    Log.e("POS IN ",""+position);
-                    holder.mNumberButton.setTextSize(TypedValue.COMPLEX_UNIT_PX, 40);
-                }
-                LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(mCustomizationOptionsBundle.getButtonSize(),mCustomizationOptionsBundle.getButtonSize());
+//                if (position == 9)
+//                {
+////                    Log.e("POS IN ",""+position);
+//                    holder.mNumberButton.setTextSize(TypedValue.COMPLEX_UNIT_PX, 40);
+//                }
+                FrameLayout.LayoutParams params = new FrameLayout.LayoutParams(mCustomizationOptionsBundle.getButtonSize(), mCustomizationOptionsBundle.getButtonSize());
                 holder.mNumberButton.setLayoutParams(params);
+            }
+        }
+    }
+
+    private void configureResetButtonHolder(ResetViewHolder holder)
+    {
+        if (holder != null)
+        {
+
+            if (mCustomizationOptionsBundle != null)
+            {
+                holder.mButtonImage.setColorFilter(mCustomizationOptionsBundle.getTextColor(), PorterDuff.Mode.SRC_ATOP);
+                LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(80, 80);
+                holder.mButtonImage.setLayoutParams(params);
             }
         }
     }
@@ -134,9 +154,9 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
                 {
                     holder.mButtonImage.setImageDrawable(mCustomizationOptionsBundle.getDeleteButtonDrawable());
                 }
-                holder.mButtonImage.setColorFilter(mCustomizationOptionsBundle.getTextColor(),PorterDuff.Mode.SRC_ATOP);
+                holder.mButtonImage.setColorFilter(mCustomizationOptionsBundle.getTextColor(), PorterDuff.Mode.SRC_ATOP);
 //                LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(mCustomizationOptionsBundle.getDeleteButtonSize(),mCustomizationOptionsBundle.getDeleteButtonSize());
-                LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(80,80);
+                LinearLayout.LayoutParams params = new LinearLayout.LayoutParams(90, 90);
 
 //                Log.e("getDeleteButtonSize ",""+mCustomizationOptionsBundle.getDeleteButtonSize());
                 holder.mButtonImage.setLayoutParams(params);
@@ -160,6 +180,10 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
         if (position == getItemCount() - 1)
         {
             return VIEW_TYPE_DELETE;
+        }
+        else if (position == getItemCount() - 3)
+        {
+            return VIEW_TYPE_RESET;
         }
         return VIEW_TYPE_NUMBER;
     }
@@ -243,28 +267,37 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
     {
         Button mNumberButton;
 
-        public NumberViewHolder(final View itemView)
+        NumberViewHolder(final View itemView)
         {
             super(itemView);
-            mNumberButton = (Button) itemView.findViewById(R.id.button);
-            mNumberButton.setOnClickListener(new View.OnClickListener()
+            mNumberButton = itemView.findViewById(R.id.button);
+            mNumberButton.setOnClickListener(v ->
             {
-                @Override
-                public void onClick(View v)
+                if (mOnNumberClickListener != null)
                 {
-                    if (mOnNumberClickListener != null)
-                    {
+                    mOnNumberClickListener.onNumberClicked((Integer) v.getTag());
+                }
+            });
+        }
+    }
 
-                        if (v.getTag().equals("reset"))
-                        {
+    public class ResetViewHolder extends RecyclerView.ViewHolder
+    {
+        LinearLayout mDeleteButton;
+        ImageView mButtonImage;
 
-                            onResetClickListener.onResetClicked();
-                        }
-                        else
-                        {
-                            mOnNumberClickListener.onNumberClicked((Integer) v.getTag());
-                        }
-                    }
+        ResetViewHolder(final View itemView)
+        {
+            super(itemView);
+            mDeleteButton = itemView.findViewById(R.id.button);
+            mButtonImage = itemView.findViewById(R.id.buttonImage);
+            mButtonImage.setImageResource(R.drawable.ic_reset);
+
+            mDeleteButton.setOnClickListener(v ->
+            {
+                if (onResetClickListener != null)
+                {
+                    onResetClickListener.onResetClicked();
                 }
             });
         }
@@ -275,40 +308,32 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
         LinearLayout mDeleteButton;
         ImageView mButtonImage;
 
-        public DeleteViewHolder(final View itemView)
+        DeleteViewHolder(final View itemView)
         {
             super(itemView);
-            mDeleteButton = (LinearLayout) itemView.findViewById(R.id.button);
-            mButtonImage = (ImageView) itemView.findViewById(R.id.buttonImage);
+            mDeleteButton =  itemView.findViewById(R.id.button);
+            mButtonImage =  itemView.findViewById(R.id.buttonImage);
 
             if (mCustomizationOptionsBundle.isShowDeleteButton() && mPinLength > 0)
             {
-                mDeleteButton.setOnClickListener(new View.OnClickListener()
+                mDeleteButton.setOnClickListener(v ->
                 {
-                    @Override
-                    public void onClick(View v)
-                    {
-                        Log.e("DELETE Adapter","setOnClickListener");
+                    Log.e("DELETE Adapter", "setOnClickListener");
 
-                        if (mOnDeleteClickListener != null)
-                        {
-                            mOnDeleteClickListener.onDeleteClicked();
-                        }
+                    if (mOnDeleteClickListener != null)
+                    {
+                        mOnDeleteClickListener.onDeleteClicked();
                     }
                 });
 
-                mDeleteButton.setOnLongClickListener(new View.OnLongClickListener()
+                mDeleteButton.setOnLongClickListener(v ->
                 {
-                    @Override
-                    public boolean onLongClick(View v)
+                    Log.e("DELETE Adapter", "setOnLongClickListener");
+                    if (mOnDeleteClickListener != null)
                     {
-                        Log.e("DELETE Adapter","setOnLongClickListener");
-                        if (mOnDeleteClickListener != null)
-                        {
-                            mOnDeleteClickListener.onDeleteLongClicked();
-                        }
-                        return true;
+                        mOnDeleteClickListener.onDeleteLongClicked();
                     }
+                    return true;
                 });
 
                 mDeleteButton.setOnTouchListener(new View.OnTouchListener()
@@ -319,13 +344,13 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
                     public boolean onTouch(View v, MotionEvent event)
                     {
 
-                        Log.e("DELETE Adapter","onTouch");
+                        Log.e("DELETE Adapter", "onTouch");
 
 
                         if (event.getAction() == MotionEvent.ACTION_DOWN)
                         {
                             mButtonImage.setColorFilter(mCustomizationOptionsBundle
-                                      .getDeleteButtonPressesColor());
+                                    .getDeleteButtonPressesColor());
                             rect = new Rect(v.getLeft(), v.getTop(), v.getRight(), v.getBottom());
                         }
                         if (event.getAction() == MotionEvent.ACTION_UP)
@@ -335,7 +360,7 @@ public class PinLockAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder
                         if (event.getAction() == MotionEvent.ACTION_MOVE)
                         {
                             if (!rect.contains(v.getLeft() + (int) event.getX(),
-                                      v.getTop() + (int) event.getY()))
+                                    v.getTop() + (int) event.getY()))
                             {
                                 mButtonImage.clearColorFilter();
                             }
